Debugging (unsuccesfully :( )
Refactored the button/lever/door/vent codes so they should behave properly now. The way I coded it feels a bit roundabout, but each other method I tried always had some small issue which made the whole thing not work, so this is unfortunately the best I can do. DON'T TOUCH IT unless you're absolutely sure you can get all the behavior working in your new method.

Unfortunately, for some godforsaken reason, this refactoring has caused all levers in the Doors & Buttons level to break, but *not* the ones in the Bait & Switch level. I've been staring at it for an hour trying to figure out what the hell the difference is, but still couldn't find the answer... anyway, I have to push my work now so we have stuff up-to-date for tomorrows meeting, but keep in mind the Doors & Buttons level is currently unfinished.