Added end screen stats
Added stats to end screen; it now displays which entrance the player started from, which exit they finished the level at, and how long it took them to beat the level (for this entrance,exit pair).
Note: this time is currently not saved anywhere (not even within the session), idk if we wanna bother with that.